+++ engine/src/main/java/dev/gamekit/components/BoxCollider.java
@@ -1,0 +1,18 @@
+package dev.gamekit.components;
+
+import dev.gamekit.core.Constants;
+import org.dyn4j.geometry.Rectangle;
+
+/** {@link BoxCollider} defines a rectangular {@link Collider} for an entity */
+public class BoxCollider extends Collider {
+  public BoxCollider(double width, double height) {
+    super(
+      new BodyAttachedFixture(
+        new Rectangle(
+          width / Constants.PIXELS_PER_METER,
+          height / Constants.PIXELS_PER_METER
+        )
+      )
+    );
+  }
+}