+++ engine/src/main/java/dev/gamekit/components/CircleCollider.java
@@ -1,0 +1,11 @@
+package dev.gamekit.components;
+
+import dev.gamekit.core.Constants;
+import org.dyn4j.geometry.Circle;
+
+/** {@link CircleCollider} defines a circular {@link Collider} for an entity */
+public class CircleCollider extends Collider {
+  public CircleCollider(double radius) {
+    super(new BodyAttachedFixture(new Circle(radius / Constants.PIXELS_PER_METER)));
+  }
+}