+++ engine/src/main/java/dev/gamekit/components/RigidBody.java
@@ -1,14 +1,17 @@
 package dev.gamekit.components;
 
-import dev.gamekit.core.*;
 import dev.gamekit.core.Component;
+import dev.gamekit.core.Constants;
+import dev.gamekit.core.Physics;
+import dev.gamekit.core.Renderer;
 import dev.gamekit.utils.Vector;
 import org.dyn4j.dynamics.Body;
-import org.dyn4j.dynamics.BodyFixture;
-import org.dyn4j.geometry.*;
-import org.dyn4j.geometry.Rectangle;
+import org.dyn4j.geometry.Mass;
+import org.dyn4j.geometry.MassType;
+import org.dyn4j.geometry.Vector2;
 
 import java.awt.*;
+import java.util.List;
 
 import static dev.gamekit.utils.Math.*;
 
@@ -19,7 +22,7 @@ import static dev.gamekit.utils.Math.*;
  * component will update the attached entity's {@link Transform} component's position and rotation
  * <p>
  * Once a {@link RigidBody} is added to an entity, position and rotation changes should be done
- * with the rigidbody component and not the {@link Transform} component
+ * with the {@link RigidBody} component and not the {@link Transform} component
  */
 public class RigidBody extends Component {
   public static boolean DEBUG_DRAW = false;
@@ -57,8 +60,10 @@ public class RigidBody extends Component {
 
   /** Sets the world position of this {@link RigidBody} */
   public void setPosition(double x, double y) {
-    double ppm = Constants.PIXELS_PER_METER;
-    body.getTransform().setTranslation(x / ppm, y / ppm);
+    body.getTransform().setTranslation(
+      x / Constants.PIXELS_PER_METER,
+      y / Constants.PIXELS_PER_METER
+    );
   }
 
   /** Sets the global rotation of this {@link RigidBody} about its center */
@@ -73,63 +78,17 @@ public class RigidBody extends Component {
    * {@code (sx, sy)} is required
    */
   public void setRotation(double deg, double rx, double ry, double sx, double sy) {
-    double ppm = Constants.PIXELS_PER_METER;
-
     body.getTransform().setRotation(0);
-    body.getTransform().setTranslation(sx / ppm, sy / ppm);
-    body.rotate(-degToRad(deg), rx / ppm, ry / ppm);
-  }
 
-  /**
-   * Creates a circle with the specified {@code radius} and attaches it to this
-   * {@link RigidBody} via a {@link BodyFixture}.
-   * <p>
-   * After creation, the new {@link Circle} and {@link BodyFixture} are passed to the
-   * {@code tuner} for configuration before they are added to this {@link RigidBody}
-   */
-  public void addCircleFixture(double radius, FixtureTuner<Circle> tuner) {
-    double ppm = Constants.PIXELS_PER_METER;
+    body.getTransform().setTranslation(
+      sx / Constants.PIXELS_PER_METER,
+      sy / Constants.PIXELS_PER_METER
+    );
 
-    Circle circle = new Circle(radius / ppm);
-    BodyFixture fx = new BodyFixture(circle);
-    tuner.tuneFixture(fx, circle);
-    body.addFixture(fx);
-    body.updateMass();
-  }
-
-  /**
-   * Creates a circle with the specified {@code radius} and attaches it to this
-   * {@link RigidBody} via a {@link BodyFixture} without additional tuning
-   * @see #addCircleFixture(double, FixtureTuner)
-   */
-  public void addCircleFixture(double radius) {
-    addCircleFixture(radius, (fx, shape) -> { });
-  }
-
-  /**
-   * Creates a rectangle with the specified {@code size} and {@code height} and attaches it to
-   * this {@link RigidBody} via a {@link BodyFixture}
-   * <p>
-   * After creation, the new {@link Rectangle} and {@link BodyFixture} are passed to the {@code
-   * tuner} for configuration before they are added to this {@link RigidBody}
-   */
-  public void addRectFixture(double width, double height, FixtureTuner<Rectangle> tuner) {
-    double ppm = Constants.PIXELS_PER_METER;
-
-    Rectangle rect = new Rectangle(width / ppm, height / ppm);
-    BodyFixture fx = new BodyFixture(rect);
-    tuner.tuneFixture(fx, rect);
-    body.addFixture(fx);
-    body.updateMass();
-  }
-
-  /**
-   * Creates a rectangle with the specified {@code size} and {@code height} and attaches it to
-   * this {@link RigidBody} via a {@link BodyFixture} without additional tuning
-   * @see #addRectFixture(double, double, FixtureTuner)
-   */
-  public void addRectFixture(double width, double height) {
-    addRectFixture(width, height, (fx, shape) -> { });
+    body.rotate(-degToRad(deg),
+      rx / Constants.PIXELS_PER_METER,
+      ry / Constants.PIXELS_PER_METER
+    );
   }
 
   /**
@@ -160,69 +119,34 @@ public class RigidBody extends Component {
 
   @Override
   protected void start() {
+    // Find all colliders and add their fixtures
+    List<Collider> colliders = entity.findComponents(Collider.class);
+    colliders.forEach(collider -> collider.fixture.addToBody(body));
+
     Physics.addBody(body);
   }
 
   @Override
   protected void update() {
-    double ppm = Constants.PIXELS_PER_METER;
     Transform tx = entity.findComponent(Transform.class);
     Vector2 center = body.getWorldCenter();
 
-    tx.setPosition(center.x * ppm, center.y * ppm);
+    tx.setPosition(center.x * Constants.PIXELS_PER_METER, center.y * Constants.PIXELS_PER_METER);
     tx.setRotation(radToDeg(body.getTransform().getRotationAngle()));
   }
 
   @Override
   protected void render() {
     if (DEBUG_DRAW) {
-      double ppm = Constants.PIXELS_PER_METER;
       Vector2 bodyCenter = body.getWorldCenter();
-      int bodyCenterX = toInt(bodyCenter.x * ppm);
-      int bodyCenterY = toInt(bodyCenter.y * ppm);
-      double bodyRotation = radToDeg(-body.getTransform().getRotationAngle());
-
-      for (BodyFixture fx : body.getFixtures()) {
-        Convex shape = fx.getShape();
-        Vector2 shapeCenter = shape.getCenter();
-        int shapeCenterX = toInt((bodyCenter.x + shapeCenter.x) * ppm);
-        int shapeCenterY = toInt((bodyCenter.y + shapeCenter.y) * ppm);
-
-        if (shape instanceof Circle circle) {
-          int radius = toInt(circle.getRadius() * ppm);
-
-          Renderer.drawCircle(shapeCenterX, shapeCenterY, radius)
-            .withColor(Color.CYAN).withRotation(bodyCenterX, bodyCenterY, bodyRotation);
-          Renderer.drawVerticalLine(shapeCenterX, shapeCenterY, shapeCenterY + radius)
-            .withRotation(bodyCenterX, bodyCenterY, bodyRotation);
-        } else if (shape instanceof Rectangle rect) {
-          int width = toInt(rect.getWidth() * ppm);
-          int height = toInt(rect.getHeight() * ppm);
-
-          Renderer.drawRect(shapeCenterX, shapeCenterY, width, height)
-            .withColor(Color.CYAN).withRotation(bodyCenterX, bodyCenterY, bodyRotation);
-          Renderer.drawVerticalLine(shapeCenterX, shapeCenterY, shapeCenterY + height / 2)
-            .withRotation(bodyCenterX, bodyCenterY, bodyRotation);
-        }
-      }
-
+      int bodyCenterX = toInt(bodyCenter.x * Constants.PIXELS_PER_METER);
+      int bodyCenterY = toInt(bodyCenter.y * Constants.PIXELS_PER_METER);
       Renderer.fillCircle(bodyCenterX, bodyCenterY, 3).withColor(Color.RED);
     }
   }
 
   @Override
   protected void dispose() {
-    Application.getInstance().runLater(
-      () -> Physics.removeBody(body)
-    );
-  }
-
-  /** Interface for an object which tunes a {@link BodyFixture} */
-  public interface FixtureTuner<S extends Convex> {
-    /**
-     * Called with the new {@link BodyFixture} and {@link Convex Shape} before adding to a
-     * {@link Body}
-     */
-    void tuneFixture(BodyFixture fixture, S shape);
+    Physics.removeBody(body);
   }
 }