+++ engine/src/main/java/dev/gamekit/core/Physics.java
@@ -1,5 +1,6 @@
 package dev.gamekit.core;
 
+import dev.gamekit.components.Collider;
 import org.dyn4j.dynamics.Body;
 import org.dyn4j.dynamics.BodyFixture;
 import org.dyn4j.world.ManifoldCollisionData;
@@ -25,17 +26,23 @@ public final class Physics {
 
         Body body1 = collision.getBody1();
         Body body2 = collision.getBody2();
-        BodyFixture fixture1 = collision.getFixture1();
-        BodyFixture fixture2 = collision.getFixture2();
+        BodyFixture fx1 = collision.getFixture1();
+        BodyFixture fx2 = collision.getFixture2();
 
         if (COLLISION_LISTENER_MAP.containsKey(body1)) {
           CollisionListener listener = COLLISION_LISTENER_MAP.get(body1);
-          listener.onCollision(body1, fixture1, body2, fixture2);
+          listener.handleCollision(
+            (Collider.BodyAttachedFixture) fx1,
+            (Collider.BodyAttachedFixture) fx2
+          );
         }
 
         if (COLLISION_LISTENER_MAP.containsKey(body2)) {
           CollisionListener listener = COLLISION_LISTENER_MAP.get(body2);
-          listener.onCollision(body2, fixture2, body1, fixture1);
+          listener.handleCollision(
+            (Collider.BodyAttachedFixture) fx2,
+            (Collider.BodyAttachedFixture) fx1
+          );
         }
 
         return true;
@@ -52,6 +59,8 @@ public final class Physics {
   static void update() {
     double elapsedTime = Constants.FRAME_INTERVAL_MS / 1000.0;
     WORLD.update(elapsedTime);
+
+    COLLISION_LISTENER_MAP.forEach((body, listener) -> listener.update());
   }
 
   /** Adds a {@link Body} to the physics world for simulation */
@@ -65,7 +74,11 @@ public final class Physics {
     WORLD.removeBody(body);
   }
 
-  /** Registers a collision listener for the specified {@link Body} */
+  /**
+   * Registers a collision listener for the specified {@link Body}.
+   * <p>
+   * If a listener is already registered for the body, it will be replaced.
+   */
   public static void addCollisionListener(Body body, CollisionListener listener) {
     COLLISION_LISTENER_MAP.put(body, listener);
   }
@@ -75,9 +88,40 @@ public final class Physics {
     COLLISION_LISTENER_MAP.remove(body, listener);
   }
 
-  /** Callback interface for physics world collisions */
-  public interface CollisionListener {
+  /** Abstract interface for handling physics collisions */
+  public static abstract class CollisionListener {
+    private final HashMap<String, Collider.BodyAttachedFixture> prevFixtureMap
+      = new HashMap<>();
+    private final HashMap<String, Collider.BodyAttachedFixture> currentFixtureMap
+      = new HashMap<>();
+
+    void update() {
+      for (String id : prevFixtureMap.keySet()) {
+        if (!currentFixtureMap.containsKey(id)) {
+          onCollisionExit(prevFixtureMap.get(id));
+        }
+      }
+
+      prevFixtureMap.clear();
+      prevFixtureMap.putAll(currentFixtureMap);
+      currentFixtureMap.clear();
+    }
+
     /** Called when a collision occurs and is passed the bodies and fixtures of the collision */
-    void onCollision(Body body1, BodyFixture fixture1, Body body2, BodyFixture fixture2);
+    void handleCollision(
+      Collider.BodyAttachedFixture ignore, Collider.BodyAttachedFixture otherFixture
+    ) {
+      if (prevFixtureMap.containsKey(otherFixture.id)) {
+        onCollisionEnter(otherFixture);
+      } else onCollisionStay(otherFixture);
+
+      currentFixtureMap.put(otherFixture.id, otherFixture);
+    }
+
+    public void onCollisionEnter(Collider.BodyAttachedFixture otherFixture) { }
+
+    public void onCollisionStay(Collider.BodyAttachedFixture otherFixture) { }
+
+    public void onCollisionExit(Collider.BodyAttachedFixture otherFixture) { }
   }
 }